Regular upmerge.

+++ src/core/thread/lowpan.hpp
@@ -91,7 +91,7 @@ public:
      * @param[in]  aLength  The size of the write buffer.
      *
      */
-    BufferWriter(uint8_t *aBuf, uint8_t aLength)
+    BufferWriter(uint8_t *aBuf, uint16_t aLength)
     {
         mWritePointer    = aBuf;
         mRemainingLength = aLength;
@@ -215,7 +215,7 @@ public:
 
 private:
     uint8_t *mWritePointer;
-    uint8_t  mRemainingLength;
+    uint16_t mRemainingLength;
 };
 
 /**
@@ -417,18 +417,18 @@ public:
      * @param[in]  aFrameLength  The length of the frame.
      *
      * @retval OT_ERROR_NONE     Mesh Header initialized successfully.
-     * @retval OT_ERROR_FAILED   Mesh header could not be initialized from @p aFrame (e.g., frame not long enough).
+     * @retval OT_ERROR_PARSE    Mesh Header could not be parsed from @p aFrame.
      *
      */
-    otError Init(const uint8_t *aFrame, uint8_t aFrameLength);
+    otError Init(const uint8_t *aFrame, uint16_t aFrameLength);
 
     /**
      * This method initializes the mesh header from a message object @p aMessage.
      *
      * @param[in]  aMessage  The message object.
      *
-     * @retval OT_ERROR_NONE     Mesh Header initialized successfully.
-     * @retval OT_ERROR_FAILED   Mesh header could not be initialized from @ aMessage(e.g., not long enough).
+     * @retval OT_ERROR_NONE   Mesh Header initialized successfully.
+     * @retval OT_ERROR_PARSE  Mesh Header could not be parsed from @p aMessage.
      *
      */
     otError Init(const Message &aMessage);
@@ -606,10 +606,10 @@ public:
      * @param[in]  aFrameLength  The length of the frame.
      *
      * @retval OT_ERROR_NONE     Fragment Header initialized successfully.
-     * @retval OT_ERROR_PARSE    Fragment header could not be initialized from @p aFrame (e.g., frame not long enough).
+     * @retval OT_ERROR_PARSE    Fragment header could not be parsed from @p aFrame.
      *
      */
-    otError Init(const uint8_t *aFrame, uint8_t aFrameLength);
+    otError Init(const uint8_t *aFrame, uint16_t aFrameLength);
 
     /**
      * This method initializes the fragment header from a message @p aMessage.
@@ -618,7 +618,7 @@ public:
      * @param[in]  aOffset       An offset into the message to read the header.
      *
      * @retval OT_ERROR_NONE     Fragment Header initialized successfully.
-     * @retval OT_ERROR_PARSE    Fragment header could not be initialized (e.g., no frag header or message too short).
+     * @retval OT_ERROR_PARSE    Fragment header could not be parsed from @p aMessage.
      *
      */
     otError Init(const Message &aMessage, uint16_t aOffset);
